// src/generated/models/Stf_alignmentsessionsModel.ts
export const Stf_alignmentsessionsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Stf_alignmentsessionsstatecode = keyof typeof Stf_alignmentsessionsstatecode;
export const Stf_alignmentsessionsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Stf_alignmentsessionsstatuscode = keyof typeof Stf_alignmentsessionsstatuscode;
export const Stf_alignmentsessionsstf_cycle = {
  1: 'Annual',
  2: 'Q1',
  3: 'Q2',
  4: 'Q3',
  5: 'Q4'
} as const;
export type Stf_alignmentsessionsstf_cycle = keyof typeof Stf_alignmentsessionsstf_cycle;
export const Stf_alignmentsessionsstf_overallverdict = {
  1: 'Sufficient',
  2: 'Partial',
  3: 'Insufficient'
} as const;
export type Stf_alignmentsessionsstf_overallverdict = keyof typeof Stf_alignmentsessionsstf_overallverdict;
export const Stf_alignmentsessionsstf_reason = {
  1: 'Insufficient Tactics or POCs',
  2: 'KPI Deficit',
  3: 'Cross Departmental Alignment'
} as const;
export type Stf_alignmentsessionsstf_reason = keyof typeof Stf_alignmentsessionsstf_reason;
export const Stf_alignmentsessionsstf_sessionstate = {
  1: 'Not Started',
  2: 'Done',
  3: 'Cancelled'
} as const;
export type Stf_alignmentsessionsstf_sessionstate = keyof typeof Stf_alignmentsessionsstf_sessionstate;
export const Stf_alignmentsessionsstf_status = {
  1: 'Open',
  2: 'In Review',
  3: 'Aligned',
  4: 'Gap Remaining',
  5: 'Closed'
} as const;
export type Stf_alignmentsessionsstf_status = keyof typeof Stf_alignmentsessionsstf_status;

export interface Stf_alignmentsessionsBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Stf_alignmentsessionsstatecode;
  statuscode?: Stf_alignmentsessionsstatuscode;
  stf_alignmentsessionid: string;
  "stf_BusinessUnit@odata.bind"?: string;
  "stf_Chair@odata.bind"?: string;
  stf_cycle: Stf_alignmentsessionsstf_cycle;
  stf_decisionnote?: string;
  "stf_Department@odata.bind"?: string;
  "stf_Facilitator@odata.bind"?: string;
  stf_fiscalyear: number;
  stf_name?: string;
  stf_overallverdict?: Stf_alignmentsessionsstf_overallverdict;
  "stf_ParentStrategy@odata.bind"?: string;
  stf_reason: Stf_alignmentsessionsstf_reason;
  "stf_Region@odata.bind"?: string;
  stf_sessiondate?: string;
  stf_sessionstate?: Stf_alignmentsessionsstf_sessionstate;
  stf_status?: Stf_alignmentsessionsstf_status;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Stf_alignmentsessions extends Stf_alignmentsessionsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  stf_businessunitname?: string;
  stf_chairname?: string;
  stf_chairyominame?: string;
  stf_cyclename?: string;
  stf_departmentname?: string;
  stf_facilitatorname?: string;
  stf_facilitatoryominame?: string;
  stf_overallverdictname?: string;
  stf_parentstrategyname?: string;
  stf_reasonname?: string;
  stf_regionname?: string;
  stf_sessionstatename?: string;
  stf_statusname?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  stf_businessunit?: object;
  _stf_businessunit_value?: string;
  stf_chair?: object;
  _stf_chair_value?: string;
  stf_department?: object;
  _stf_department_value?: string;
  stf_facilitator?: object;
  _stf_facilitator_value?: string;
  stf_parentstrategy?: object;
  _stf_parentstrategy_value?: string;
  stf_region?: object;
  _stf_region_value?: string;
}